+++ src/components/form/FormElementWrapper.tsx
@@ -1,23 +1,24 @@
 import type { HTMLAttributes, ReactNode } from 'react'
-import { useEffect } from 'react'
-import { useState } from 'react'
 import { useId } from 'react'
 import { clsx } from 'clsx'
 import type { BagFunction } from '@/src/utils/bagFunctions'
 import type { LabelProps } from '@/src/components/user-action/Label'
 import { Label } from '@/src/components/user-action/Label'
+import { useOverwritableState } from '@/src/hooks/useOverwritableState'
 
 type ErrorShowBehaviour = 'always' | 'whenTouched'
 
-type FormElementWrapperBag = {
-  'touched': boolean,
-  'onTouched': (isTouched?: boolean) => void,
-  'invalid': boolean,
-  'required': boolean,
-  'disabled': boolean,
-  'aria-describedby': string,
-  'aria-labelledby': string,
-  'id': string,
+type FormElementHTMLProperties = Pick<HTMLAttributes<HTMLElement>, 'id' | 'aria-labelledby' | 'aria-describedby' | 'aria-disabled' | 'aria-invalid' | 'aria-errormessage' | 'aria-required'>
+
+export type FormElementStateProperties = {
+  invalid: boolean,
+  disabled: boolean,
+}
+
+type FormElementWrapperBag = FormElementHTMLProperties & FormElementStateProperties & {
+  required: boolean,
+  isShowingError: boolean,
+  setIsShowingError: (isShowingError?: boolean) => void,
 }
 
 export type FormElementWrapperProps = {
@@ -25,7 +26,8 @@ export type FormElementWrapperProps = {
   id?: string,
   required?: boolean,
   disabled?: boolean,
-  isTouched?: boolean,
+  isShowingError?: boolean,
+  onIsShowingError?: (isTouched: boolean) => void,
   label?: ReactNode,
   labelProps?: Omit<LabelProps, 'children'>,
   errorShowBehaviour?: ErrorShowBehaviour,
@@ -41,7 +43,8 @@ export const FormElementWrapper = ({
                                      id,
                                      required = false,
                                      disabled = false,
-                                     isTouched: initialIsTouched = false,
+                                     isShowingError: initialIsShowingError = false,
+                                     onIsShowingError,
                                      label,
                                      labelProps,
                                      errorShowBehaviour = 'whenTouched',
@@ -51,17 +54,17 @@ export const FormElementWrapper = ({
                                      descriptionProps,
                                      containerClassName,
                                    }: FormElementWrapperProps) => {
-  const [touched, setTouched] = useState(initialIsTouched)
+  const [touched, setTouched] = useOverwritableState(initialIsShowingError, onIsShowingError)
   const generatedId = useId()
   const usedId = id ?? generatedId
 
-  useEffect(() => {
-    setTouched(initialIsTouched)
-  }, [initialIsTouched])
+  const errorId = `${usedId}-description`
+  const labelId = `${usedId}-label`
+  const descriptionId = `${usedId}-description`
 
   const describedBy: string = [
-    description ? `${usedId}-description` : undefined,
-    error ? `${usedId}-error` : undefined,
+    description ? descriptionId : undefined,
+    error ? errorId : undefined,
   ].filter(Boolean).join(' ')
 
   const labelledBy = label ? `${usedId}-label` : undefined
@@ -69,32 +72,57 @@ export const FormElementWrapper = ({
   const isShowingError = errorShowBehaviour === 'always' || (touched)
 
   const bag: FormElementWrapperBag = {
-    'touched': touched,
-    'onTouched': (isTouched) => setTouched(isTouched ?? true),
+    'isShowingError': touched,
+    'setIsShowingError': (isTouched) => {
+      const newValue = isTouched ?? true
+      setTouched(newValue)
+      onIsShowingError?.(isTouched)
+    },
     'disabled': disabled,
     'invalid': !!error && isShowingError,
     'required': required,
     'id': usedId,
+    'aria-required': required,
+    'aria-disabled': disabled,
     'aria-describedby': describedBy,
-    'aria-labelledby': labelledBy
+    'aria-labelledby': labelledBy,
+    'aria-invalid': !!error,
+    'aria-errormessage': error ? errorId : undefined,
   }
 
   return (
     <div className={clsx('relative flex flex-col gap-y-1', containerClassName)}>
       {label && (
-        <Label htmlFor={usedId} size="lg" className={clsx('flex-row-1 items-start', labelProps?.className)}>
+        <Label
+          {...labelProps}
+          id={labelId}
+          htmlFor={usedId}
+          size="lg"
+          className={clsx('flex-row-1 items-start', labelProps?.className)}
+        >
           {label}
           {required && <div role="none" className="bg-primary w-2 h-2 rounded-full"/>}
         </Label>
       )}
       {description && (
-        <p {...descriptionProps} className={clsx('text-description text-sm', descriptionProps?.className)}>
+        <p
+          {...descriptionProps}
+          id={descriptionId}
+          className={clsx('text-description text-sm', descriptionProps?.className)}
+        >
           {description}
         </p>
       )}
       {children(bag)}
       {error && isShowingError && (
-        <p {...errorProps} role="alert" className={clsx('absolute top-[calc(100%_+_0.25rem)] left-0 text-negative text-sm font-medium', errorProps)}>
+        <p
+          {...errorProps}
+          id={errorId}
+          className={clsx('absolute top-[calc(100%_+_0.25rem)] left-0 text-negative text-sm font-medium', errorProps?.className)}
+          role="alert"
+          aria-hidden={!error}
+          aria-live="polite"
+        >
           {error}
         </p>
       )}