+++ src/components/icons-and-geometry/index.ts
@@ -1,5 +1,0 @@
-export * from './Avatar'
-export * from './Circle'
-export * from './HelpwaveLogo'
-export * from './Ring'
-export * from './Tag'