+++ src/components/navigation/index.ts
@@ -1,4 +1,0 @@
-export * from './BreadCrumb'
-export * from './Navigation'
-export * from './Pagination'
-export * from './StepperBar'