+++ src/components/table/Table.tsx
@@ -27,7 +27,6 @@ import {
 import { range } from '@/src/utils/array'
 import { Scrollbars } from 'react-custom-scrollbars-2'
 import { clamp } from '@/src/utils/math'
-import { noop } from '@/src/utils/noop'
 import { TableCell } from '@/src/components/table/TableCell'
 import { TableFilters } from '@/src/components/table/Filter'
 import { useResizeCallbackWrapper } from '@/src/hooks/useResizeCallbackWrapper'
@@ -37,6 +36,7 @@ import { TableFilterButton } from '@/src/components/table/TableFilterButton'
 import { FillerRowElement } from '@/src/components/table/FillerRowElement'
 import { Pagination } from '@/src/components/navigation/Pagination'
 import { Checkbox } from '@/src/components/user-action/Checkbox'
+import { useOverwritableState } from '@/src/hooks/useOverwritableState'
 
 declare module '@tanstack/react-table' {
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
@@ -76,7 +76,7 @@ export const Table = <T, >({
                              data,
                              fillerRow,
                              initialState,
-                             onRowClick = noop,
+                             onRowClick,
                              className,
                              tableClassName,
                              defaultColumn,
@@ -397,7 +397,7 @@ export const Table = <T, >({
           <tbody>
           {table.getRowModel().rows.map(row => {
             return (
-              <tr key={row.id} onClick={() => onRowClick(row, table)} className={table.options.meta?.bodyRowClassName}>
+              <tr key={row.id} onClick={() => onRowClick?.(row, table)} className={table.options.meta?.bodyRowClassName}>
                 {row.getVisibleCells().map(cell => {
                   return (
                     <td key={cell.id}>
@@ -442,11 +442,7 @@ export const Table = <T, >({
 export type TableUncontrolledProps<T> = TableProps<T>
 
 export const TableUncontrolled = <T, >({ data, ...props }: TableUncontrolledProps<T>) => {
-  const [usedDate, setUsedData] = useState<T[]>(data)
-
-  useEffect(() => {
-    setUsedData(data)
-  }, [data])
+  const [usedDate] = useOverwritableState<T[]>(data)
 
   return (
     <Table
@@ -470,7 +466,7 @@ export const TableWithSelection = <T, >({
                                           rowSelection,
                                           disableClickRowClickSelection = false,
                                           selectionRowId = 'selection',
-                                          onRowClick = noop,
+                                          onRowClick,
                                           meta,
                                           ...props
                                         }: TableWithSelectionProps<T>) => {
@@ -526,7 +522,7 @@ export const TableWithSelection = <T, >({
         if (!disableClickRowClickSelection) {
           row.toggleSelected()
         }
-        onRowClick(row, table)
+        onRowClick?.(row, table)
       }}
       meta={{
         ...meta,