+++ src/hooks/useLocalStorage.ts
@@ -6,15 +6,28 @@ import { LocalStorageService } from '@/src/utils/storage'
 import { resolveSetState } from '@/src/utils/resolveSetState'
 
 type SetValue<T> = Dispatch<SetStateAction<T>>
-export const useLocalStorage = <T>(key: string, initValue: T): [T, SetValue<T>] => {
+
+type UseLocalStorageResult<T> = {
+  value: T,
+  setValue: SetValue<T>,
+  deleteValue: () => void,
+}
+
+/**
+ * @param key Key under which to save the data
+ * @param backupValue Used if the storage is unavailable or no value is present
+ *
+ * The backup value will never be saved to the storage unless you explicitly
+ */
+export const useLocalStorage = <T>(key: string, backupValue: T): UseLocalStorageResult<T> => {
   const get = useCallback((): T => {
     if (typeof window === 'undefined') {
-      return initValue
+      return backupValue
     }
     const storageService = new LocalStorageService()
     const value = storageService.get<T>(key)
-    return value || initValue
-  }, [initValue, key])
+    return value || backupValue
+  }, [backupValue, key])
 
   const [storedValue, setStoredValue] = useState<T>(get)
 
@@ -26,5 +39,11 @@ export const useLocalStorage = <T>(key: string, initValue: T): [T, SetValue<T>] 
     setStoredValue(newValue)
   }, [storedValue, setStoredValue, key])
 
-  return [storedValue, setValue]
+  const deleteValue = () => {
+    const storageService = new LocalStorageService()
+    storageService.delete(key)
+    setStoredValue(backupValue)
+  }
+
+  return { value: storedValue, setValue, deleteValue }
 }