+++ src/index.ts
@@ -1,6 +1,0 @@
-export * from './coloring'
-export * from './components'
-export * from './hooks'
-export * from './localization'
-export * from './theming'
-export * from './utils'